Menu window (#13)
* Added DAB window to launch client and server

Added the ability to launch a window and launch the client and the
server. Changed the client to take command arguments.

* Added ignore for DABExec.java

NOTICE: CHANGE DABExec.java to your build path or Dab.java will not work

* Clean Up

Cleaned up a lot of the code and documented it

* Updated player picture

Refactored the player a bit

* Refactored how the client and server run

Made it so the client and server are run on two seperate threads inside
this single program. Game can now be launch entirely from the
DestroyAllBaddies.java file.

* Simplified to one file

* One last thing

Went through and cleaned things up one last time

* Added Icon

Added Icon to DAB